Extract class generation code out of TypeMap (#535)
